HW08 add MealUIController and meals.js add filter for meal and user enabled fucn for futer Option delete meal not worked(i don't know what happens) Add disable filter button

+++ src/main/webapp/resources/js/topjava.common.js
@@ -1,6 +1,7 @@
-var form;
+var form,context;
 
 function makeEditable() {
+    context=ctx;
     form = $('#detailsForm');
     $(".delete").click(function () {
         if (confirm('Are you sure?')) {
@@ -22,19 +23,32 @@ function add() {
 }
 
 function deleteRow(id) {
-    $.ajax({
-        url: ctx.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTable();
-        successNoty("Deleted");
-    });
+    if (confirm('Are you sure?')) {
+        $.ajax({
+            url: context.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            context.updateTable();
+            successNoty("Deleted");
+        });
+    }
+
+    // $.ajax({
+    //     url: ctx.ajaxUrl + id,
+    //     type: "DELETE"
+    // }).done(function () {
+    //     updateTable();
+    //     successNoty("Deleted");
+    // });
 }
 
-function updateTable() {
-    $.get(ctx.ajaxUrl, function (data) {
-        ctx.datatableApi.clear().rows.add(data).draw();
-    });
+// function updateTable() {
+//     $.get(ctx.ajaxUrl, function (data) {
+//         ctx.datatableApi.clear().rows.add(data).draw();
+//     });
+// }
+function updateTableByData(data) {
+    ctx.datatableApi.clear().rows.add(data).draw();
 }
 
 function save() {
@@ -44,7 +58,7 @@ function save() {
         data: form.serialize()
     }).done(function () {
         $("#editRow").modal("hide");
-        updateTable();
+        ctx.updateTable();
         successNoty("Saved");
     });
 }
